Frontend/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.14.0
qdrant-client>=1.5.0
numpy>=1.24.0
python-dotenv>=1.0.0
tenacity>=8.2.0
loguru>=0.7.0
requests
pydantic